正确终止、定期报告
