+++ settings.ini
@@ -7,10 +7,6 @@
 # 热搜->搜索->话题->内容->评论->转发->用户
 multi_process      : 0
 
-# 是否开启报错邮件提示
-# 若开启这项，需要首先安装并配置 py_reminder: https://github.com/Wenzhi-Ding/py_reminder
-monitor            : 0
-
 # 用于决定是否开启相关任务
 get_trending       : 0
 get_search         : 1
@@ -39,3 +35,10 @@ end_time           : 2021-01-01-0
 [Break]
 # 标识需要将未完成任务打散为几份
 n_worker           : 4
+
+[Email]
+# 是否开启报错邮件提示
+# 若开启这项，需要首先安装并配置 py_reminder: https://github.com/Wenzhi-Ding/py_reminder
+monitor            : 1
+# 发邮件的间隔时间，以分钟为单位
+interval           : 240
