+++ test/email_summary.py
@@ -1,0 +1,4 @@
+from summary import email, summary
+
+if __name__ == '__main__':
+    email(summary())
